resync packages
