+++ open_guidelines/.easignore
@@ -1,0 +1,9 @@
+/node_modules
+/.expo
+/android
+/ios
+/build
+/dist
+
+*.log
+.DS_Store
