+++ open_guidelines/android/.gitignore
@@ -12,5 +12,8 @@ local.properties
 *.hprof
 .cxx/
 
+# generated inline modules
+app/src/main/java/inline/
+
 # Bundle artifacts
 *.jsbundle
