+++ open_guidelines/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">open_guidelines</string>
+  <string name="app_name">Open Guidelines</string>
 </resources>
