+++ open_guidelines/package.json
@@ -9,10 +9,10 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@expo/cli": "^56.1.12",
+    "@expo/cli": "^0.0.0",
     "@react-navigation/native": "^6.1.17",
     "@react-navigation/native-stack": "^6.9.26",
-    "expo": "^56.0.5",
+    "expo": "^56.0.11",
     "expo-status-bar": "~56.0.4",
     "react": "19.2.3",
     "react-dom": "19.2.3",
